BUG: `OrdinationResults` is now importable from `skbio` and `skbio.stats.ordination` and correctly linked from the documentation
Fixes #1205.
