+++ skbio/__init__.py
@@ -17,7 +17,7 @@ from skbio.stats.distance import DistanceMatrix
 from skbio.alignment import local_pairwise_align_ssw, TabularMSA
 from skbio.tree import TreeNode, nj
 from skbio.io import read, write
-from skbio._base import OrdinationResults
+from skbio.stats.ordination import OrdinationResults
 
 
 __all__ = ['Sequence', 'DNA', 'RNA', 'Protein', 'GeneticCode',
