+++ skbio/stats/ordination/__init__.py
@@ -23,7 +23,16 @@ Functions
    corr
    scale
    svd_rank
+   e_matrix
+   f_matrix
 
+Classes
+-------
+
+.. autosummary::
+   :toctree: generated/
+
+   OrdinationResults
 
 Examples
 --------
@@ -118,9 +127,10 @@ from ._redundancy_analysis import rda
 from ._correspondence_analysis import ca
 from ._canonical_correspondence_analysis import cca
 from ._principal_coordinate_analysis import pcoa
+from ._ordination_results import OrdinationResults
 from ._utils import (mean_and_std, scale, svd_rank, corr, e_matrix, f_matrix)
 
-__all__ = ['ca', 'rda', 'cca', 'pcoa',
+__all__ = ['ca', 'rda', 'cca', 'pcoa', 'OrdinationResults',
            'mean_and_std', 'scale', 'svd_rank', 'corr',
            'e_matrix', 'f_matrix']
